Move tests to ZeroCore
