+++ test/Abp.ZeroCore.SampleApp/EntityFramework/SampleAppDbContext.cs
@@ -1,5 +1,6 @@
 ﻿using Abp.Zero.EntityFrameworkCore;
 using Abp.ZeroCore.SampleApp.Core;
+using Abp.ZeroCore.SampleApp.Core.EntityHistory;
 using Microsoft.EntityFrameworkCore;
 using Abp.IdentityServer4;
 
@@ -9,6 +10,10 @@ namespace Abp.ZeroCore.SampleApp.EntityFramework
     public class SampleAppDbContext : AbpZeroDbContext<Tenant, Role, User, SampleAppDbContext>, IAbpPersistedGrantDbContext
     {
         public DbSet<PersistedGrantEntity> PersistedGrants { get; set; }
+
+        public DbSet<Blog> Blogs { get; set; }
+
+        public DbSet<Post> Posts { get; set; }
 
         public SampleAppDbContext(DbContextOptions<SampleAppDbContext> options) 
             : base(options)
